# Conflicts:
#	Gruntfile.js
#	assets/dev/scss/frontend/_widgets-global.scss
#	assets/dev/scss/frontend/_widgets.scss
#	assets/dev/scss/frontend/frontend-lite.scss
#	assets/dev/scss/frontend/general/_base-tabs.scss
#	assets/dev/scss/frontend/general/_icon.scss
#	assets/dev/scss/frontend/widgets/icon.scss
#	core/base/document.php
#	core/base/elements-iteration-actions/assets.php
#	core/experiments/manager.php
#	includes/base/widget-base.php
#	includes/frontend.php
#	includes/plugin.php

+++ core/common/modules/connect/apps/base-app.php
@@ -386,7 +386,7 @@ abstract class Base_App {
 			$body = (object) $body;
 
 			$message = isset( $body->message ) ? $body->message : wp_remote_retrieve_response_message( $response );
-			$code = isset( $body->code ) ? $body->code : $response_code;
+			$code = (int) ( isset( $body->code ) ? $body->code : $response_code );
 
 			if ( 401 === $code ) {
 				$this->delete();
